third_party/blink/web_tests/platform/mac/external/wpt/editing/run/superscript-expected.txt

This is a testharness.js-based test.
Found 265 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["superscript",""]] "foo[]bar" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "foo[]bar" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p> <p>bar]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sup>foo</sup></p> <p><sup>bar</sup></p>" but got "<p><span style=\\"vertical-align:super\\">foo</span></p> <p><span style=\\"vertical-align:super\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandState("stylewithcss") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo</p> <p>bar]</p>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<span>[foo</span> <span>bar]</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sup><span>foo</span> <span>bar</span></sup>" but got "<span style=\\"vertical-align:super\\"><span>foo</span> <span>bar</span></span>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<span>[foo</span> <span>bar]</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sup>foo</sup></p><p> <sup><span>bar</span></sup> </p><p><sup>baz</sup></p>" but got "<p><span style=\\"vertical-align:super\\">foo</span></p><p> <span style=\\"vertical-align:super\\"><span>bar</span> </span></p><p><span style=\\"vertical-align:super\\">baz</span></p>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sup>foo</sup></p><p> <sup><span>bar</span></sup> </p><p><sup>baz</sup></p>" but got "<p><sup>foo</sup></p><p> <sup><span>bar</span> </sup></p><p><sup>baz</sup></p>"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo</p><p> <span>bar</span> </p><p>baz]</p>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo<p><br><p>bar]" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><sup>foo</sup></p><p><sup><br></sup></p><p><sup>bar</sup></p>" but got "<p><span style=\\"vertical-align:super\\">foo</span></p><p><span style=\\"vertical-align:super\\"><br></span></p><p><span style=\\"vertical-align:super\\">bar</span></p>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<p>[foo<p><br><p>bar]" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "<b>foo[]bar</b>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "<b>foo[]bar</b>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "<i>foo[]bar</i>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "<i>foo[]bar</i>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "<span>foo</span>{}<span>bar</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "<span>foo[</span><span>]bar</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar]baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<span style=\\"vertical-align:super\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar]baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar]baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[bar]baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[bar]baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar<b>baz]qoz</b>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup><b><sup>baz</sup>qoz</b>quz" but got "foo<span style=\\"vertical-align:super\\">bar</span><b><span style=\\"vertical-align:super\\">baz</span>qoz</b>quz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[bar<b>baz]qoz</b>quz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar<i>baz]qoz</i>quz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup><i><sup>baz</sup>qoz</i>quz" but got "foo<span style=\\"vertical-align:super\\">bar</span><i><span style=\\"vertical-align:super\\">baz</span>qoz</i>quz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[bar<i>baz]qoz</i>quz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<p><p> <p>foo</p>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p></p><p> </p><p><sup>foo</sup></p>" but got "<p></p><p> </p><p><span style=\\"vertical-align:super\\">foo</span></p>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "{<p><p> <p>foo</p>}" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td>b<sup>a</sup>r</td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td>b<span style=\\"vertical-align:super\\">a</span>r</td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody><tr><td>foo<td>b[a]r<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>foo</td><td><sup>bar</sup></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td>foo</td><td><span style=\\"vertical-align:super\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody><tr data-start=1 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sup>foo</sup></td><td><sup>bar</sup></td><td>baz</td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:super\\">foo</span></td><td><span style=\\"vertical-align:super\\">bar</span></td><td>baz</td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody><tr data-start=0 data-end=2><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sup>foo</sup></td><td><sup>bar</sup></td><td><sup>baz</sup></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:super\\">foo</span></td><td><span style=\\"vertical-align:super\\">bar</span></td><td><span style=\\"vertical-align:super\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table><tbody data-start=0 data-end=1><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sup>foo</sup></td><td><sup>bar</sup></td><td><sup>baz</sup></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:super\\">foo</span></td><td><span style=\\"vertical-align:super\\">bar</span></td><td><span style=\\"vertical-align:super\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "<table data-start=0 data-end=1><tbody><tr><td>foo<td>bar<td>baz</table>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td><sup>foo</sup></td><td><sup>bar</sup></td><td><sup>baz</sup></td></tr></tbody></table>" but got "<table><tbody><tr><td><span style=\\"vertical-align:super\\">foo</span></td><td><span style=\\"vertical-align:super\\">bar</span></td><td><span style=\\"vertical-align:super\\">baz</span></td></tr></tbody></table>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "{<table><tr><td>foo<td>bar<td>baz</table>}" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>[bar]</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<span style=\\"vertical-align:super\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>[bar]</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b[a]r</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sub>b</sub><span style=\\"vertical-align:super\\">a</span><sub>r</sub>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b[a]r</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>b[a]r</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<span style=\\"vertical-align:super\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:sub>[bar]</span>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:super>[bar]</span>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<span style=\\"vertical-align:super\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sub>[bar]</sub></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sub style=\\"vertical-align:sub\\">b</sub><span style=\\"vertical-align:super\\">a</span><sub style=\\"vertical-align:sub\\">r</sub>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sub style=\\"vertical-align:sub\\">b</sub><sup>a</sup><sub style=\\"vertical-align:sub\\">r</sub>baz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sub>b[a]r</sub></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">b</span><span style=\\"vertical-align:super\\">a</span><span style=\\"vertical-align:sub\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sub style=\\"vertical-align:sub\\">b</sub><sup>a</sup><sub>r</sub>baz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sub>[a]</sub>r</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sup>[bar]</sup></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup>a<sup>r</sup>baz" but got "foo<sup style=\\"vertical-align:super\\">b</sup>a<sup style=\\"vertical-align:super\\">r</sup>baz"
[FAIL] [["superscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo<sup><sup>b[a]r</sup></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>b</sup>a<sup>r</sup>baz" but got "foo<sup style=\\"vertical-align:super\\">b</sup>a<sup>r</sup>baz"
[FAIL] [["superscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo<sup>b<sup>[a]</sup>r</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foobarbaz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>[bar]</sup></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<sup style=\\"vertical-align:sub\\">b</sup>a<sup style=\\"vertical-align:sub\\">r</sup>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<sup style=\\"vertical-align:sub\\">b</sup>a<sup style=\\"vertical-align:sub\\">r</sup>baz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub><sup>b[a]r</sup></sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<span style=\\"vertical-align:sub\\">b</span>a<span style=\\"vertical-align:sub\\">r</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sub>b</sub>a<sub>r</sub>baz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sub>b<sup>[a]</sup>r</sub>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>[bar]</sub></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sup><sub>b</sub>a<sub>r</sub></sup>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sub>b</sub><sup>a</sup><sub>r</sub>baz" but got "foo<sup><sub>b</sub>a<sub>r</sub></sup>baz"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup><sub>b[a]r</sub></sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>b<sub>[a]</sub>r</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "fo[o<sup>b]ar</sup>baz" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""]] "fo[o<sup>b]ar</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "fo[o<sup>b]ar</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>b]az" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>barb</sup>az" but got "foo<sup>ba</sup>rbaz"
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>b]az" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>b]az" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>b]az" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>b]az" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>b]az" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "fo[o<sup>bar</sup>b]az" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""]] "fo[o<sup>bar</sup>b]az" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "fo[o<sup>bar</sup>b]az" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo[<sup>b]ar</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo[<sup>b]ar</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>]baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "foo<sup>ba[r</sup>]baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[<sup>bar</sup>]baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo[<sup>bar</sup>]baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[<sup>bar</sup>]baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo[<sup>bar</sup>]baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>[bar]</sup>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo{<sup>bar</sup>}baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo{<sup>bar</sup>}baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo{<sup>bar</sup>}baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo{<sup>bar</sup>}baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["superscript",""]] "<sup>fo[o</sup><sub>b]ar</sub>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sup>foob</sup><sub>ar</sub>" but got "<sup>fo</sup>ob<sub>ar</sub>"
[FAIL] [["superscript",""]] "<sup>fo[o</sup><sub>b]ar</sub>" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""]] "<sup>fo[o</sup><sub>b]ar</sub>" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["superscript",""]] "<sup>fo[o</sup><sub>b]ar</sub>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["superscript",""]] "<sup>fo[o</sup><sub>b]ar</sub>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["superscript",""]] "<sup>fo[o</sup><sub>b]ar</sub>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sup>foob</sup><span style=\\"vertical-align:super\\">ar</span>" but got "<sup>fo</sup>ob<span style=\\"vertical-align:super\\">ar</span>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sup>foob</sup><span style=\\"vertical-align:super\\">ar</span>" but got "<sup>fo</sup>ob<span style=\\"vertical-align:super\\">ar</span>"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:super>b]ar</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sup>foob</sup><span style=\\"vertical-align:bottom\\">ar</span>" but got "<sup>fo</sup>ob<span style=\\"vertical-align:bottom\\">ar</span>"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","true"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<sup>foob</sup><span style=\\"vertical-align:bottom\\">ar</span>" but got "<sup>fo</sup>ob<span style=\\"vertical-align:bottom\\">ar</span>"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandIndeterm("superscript") before
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandState("superscript") before
  assert_equals: Wrong result returned expected false but got true
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandState("superscript") after
  assert_equals: Wrong result returned expected true but got false
[FAIL] [["stylewithcss","false"],["superscript",""]] "<sup>fo[o</sup><span style=vertical-align:bottom>b]ar</span>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:bottom>[bar]</span>baz" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "foo<sup>bar</sup>baz" but got "foo<span style=\\"vertical-align:super\\">bar</span>baz"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:bottom>[bar]</span>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<span style=vertical-align:bottom>[bar]</span>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:bottom>[bar]</span>baz" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<span style=vertical-align:bottom>[bar]</span>baz" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>[bar]<br></sup>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","true"],["superscript",""]] "foo<sup>[bar]<br></sup>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>[bar]<br></sup>" queryCommandValue("superscript") before
  assert_equals: Wrong result returned expected "" but got "true"
[FAIL] [["stylewithcss","false"],["superscript",""]] "foo<sup>[bar]<br></sup>" queryCommandValue("superscript") after
  assert_equals: Wrong result returned expected "" but got "false"
Harness: the test ran to completion.
